requirements.txt
playwright>=1.49.0
google-cloud-aiplatform>=1.38.1
google-cloud-vision>=3.5.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-genai
pandas>=2.1.4
openpyxl>=3.1.2
python-dotenv>=1.0.0
asyncpg>=0.29.0
async_timeout>=4.0.3
flask
fastapi
uvicorn
httpx
